Fix kafka topic-enforcer crash in continous mode if an exception occurs
Prior to this change, topic-enforcer would crash if an error is encountered
in one of the runs in continous mode. Thats not what we want, the expected
ux is to track the error but continue to be alive and maintain continous
enforcement.

To improve testing, this change refactors cli layers.