Keystore as-a-library usability improvements
Rather than making sure you pass in the right certificates to the input
streams, now you just pass in an instance of KeystoreConfig and it does
the right thing for you when creating the keystore or truststore. However,
also allows access to underlying methods to make the stores yourself, if
you prefer.

Also, adds support for reading the Issuing CA certificate from the CA Chain,
rather than having a separate certificate. This is a common way certificates
are handed out to clients, so also improves usability dramatically (rather
than messing with manually editing certs).