Add a utility class for parsing multi cluster topic config
Motivation:

Kafka topic enforcer accepts a config file that is tied to a cluster. This
means you could have N config files for N clusters. Generally, there is a lot of
overlap in these configs, for instance, production and staging configs only
differ in properties such as partition count, replication factor, etc.

It'd be operationally simple and less error-prone to reduce the number
of configs, this change is a step in that direction.