Reduce logging verbosity for some common messages
We already track the issues via metrics, so the logging does not need to be quite so
verbose - listing every single topic/partition that can have a common error. This
can become particularly verbose when dealing with hundreds of consumers. Instead
this turns down the log levels so real errors can be more easily seen or are
summarized across all partitions (i.e. no partition has an end offset for the
consumer group, rather than listing potentially thousands of partitions).