Password generator should not include slashes
Ends up not playing nice when loading into strings in Java as they need to
be escaped. There are plenty of other 'good' special characters left to use,
so not a big risk to remove these in terms of randomness.

Test needs to be fixed because the range of characters to pick from is different,
leading to different items being selected from the list.