Add a metric to track kafka record timestamp type
Freshness tracker uses kafka record timestamp to determine
when the event may have entered our system. This timestamp can
be set in two ways

(a) A producer can set it, in which case the timestamp type is createtime
(b) Broker can set it, in which case the timestamp type is logappendtime

An edge case for freshness tracker is when producer sets the timestamp and
producer's clock is out of sync, in such case, freshness tracker will be
mislead into detecting a lag since it uses current system time (tracker's).

This change adds a new metric to track timestamp types. It doesn't do much
to avoid the false detection, but provides visibility to debug the issue further.