Burrow 404 stack traces are omitted
They end up clogging the logs as often due to burrow configs consumers
can be returned but not have any lag information available. This can
cause the logs to be filled with useless stack traces of REST 404 errors
from the burrow lag queries. Instead, we just log the one-line message
containing the Response information - everything that is actually useful
in debugging a 404.