Adding LOG.debug statements for calculating consumer freshness
This gives consumers a flag to toggle - turning on DEBUG level logging - that
though it is verbose, will allow them to see what is actually happening
under the hood in the freshness tracker. This can be very useful for debugging
'weird' results from freshness computations.

This also meant turning down the Burrow GET request logging from DEBUG to TRACE
as we really dont need to see the requests in the logs at DEBUG when we are getting
that same response information in ConsumerFreshness (albeit slightly massaged).