Improve freshness verbosity (#60)
* Burrow 404 stack traces are omitted

They end up clogging the logs as often due to burrow configs consumers
can be returned but not have any lag information available. This can
cause the logs to be filled with useless stack traces of REST 404 errors
from the burrow lag queries. Instead, we just log the one-line message
containing the Response information - everything that is actually useful
in debugging a 404.

* Adding LOG.debug statements for calculating consumer freshness

This gives consumers a flag to toggle - turning on DEBUG level logging - that
though it is verbose, will allow them to see what is actually happening
under the hood in the freshness tracker. This can be very useful for debugging
'weird' results from freshness computations.

This also meant turning down the Burrow GET request logging from DEBUG to TRACE
as we really dont need to see the requests in the logs at DEBUG when we are getting
that same response information in ConsumerFreshness (albeit slightly massaged).