Burrow does not log unsuccessful requests (#61)
That is, requests that complete but have a non-200 REST response code
are not logged at ERROR along with a stack trace. We already log the
error in the caller (ConsumerFreshness) and these burrow errors are
common enough and failed REST responses dont need full stack traces
to determine the eerror - in fact, the stack trace can often just add
noise to the log.