Upgrade Kafka and zookeeper dependencies to newer versions (#63)
Upgrade Kafka dependencies to version 2.8.1, which is the latest release
in the 2.x.x series. This release depends on Zookeeper 3.5.x, so update
that dependency to the latest version as well.

This upgrade allows using newer Kafka features. One such feature that has
an open ticket is to enforce quotas using the admin client instead of the
old zookeeper client (#55).

Co-authored-by: Rohan Agarwal <rohagarwal@tesla.com>