Update to Java version 17 (#69)
Update the following dependencies to newer versions so they work with
Java 17:
* mockito
* apache commons-lang3
* zookeeper
* jackson

Some tests were updated to work with newer version of mockito.

Update Bazel to version 6.5.0 and pull in new version of rules_java to
use the latest JDK for compilation. Build source code targeting Java 17
bytecode, and run tests with Java 17 as well.