Use double instead of integers for Kafka quota values. (#75)
This is to be consistent with Kafka's settings. Some tests are explicitly using integers in Kafka config to test the backward compatibility.

Note that this change is backward compatible. However, this commit is NOT rollback-safe, as new quota values will be set with decimals such as "10.0" in ZooKeeper even if the quota is set as integers in the config file, and if we rollback to older version of enforcer, it will fail at these decimal numbers.