Implement Kafka Admin Client based Quota Service. (#76)
We are implementing a Quota Service that does not rely on ZooKeeper following Kafka's documentation that can be found at https://kafka.apache.org/28/javadoc/org/apache/kafka/clients/admin/Admin.html. Unit tests are also included for this newly created Quota Service.

Our intention is to update and replace the current ZooKeeper based Quota Service to resolve #55. For now, we are only adding the new implementation without using it in the enforcer to avoid breaking existing use cases, while we are still adding new features.